+++ CHANGELOG.md
@@ -1,16 +1,31 @@
 # Changelog
 
-## v1.0.2
+## v1.2.0 - TBD
+
+* Add option to disable script hover
+* Update dependencies
+
+## v1.1.0 - 2022-07-08
+
+* Add [Welcome content](https://code.visualstudio.com/api/extension-guides/tree-view#welcome-content) to CommandBox Scripts view.
+* Add Workspace Trust support
+* Improve tree view
+* Use URI instead of file path where possible.
+* Replace usages of Node `fs` and `path` with VS Code native APIs where possible.
+* Update dependencies
+* ESLint fixes
+
+## v1.0.2 - 2020-04-08
 
 * Bump dependencies
 * Replace bundled icons with codicon
 * Remove workaround for hover run script bug
 
-## v1.0.1
+## v1.0.1 - 2020-03-19
 
 * Bump dependencies
 * Clarify that VS Code tasks are what is being used
 
-## v1.0.0
+## v1.0.0 - 2020-03-18
 
 * Initial release